Move the fixtures back to test_drop_Counters
